requirements.txt
pandas>=2.0.0
streamlit>=1.28.0
google-generativeai>=0.8.0
google-genai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
requests>=2.28.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
tabulate>=0.9.0
